Move file to root level
