+++ framework/src/modules/chain/init_steps/init_modules.js
@@ -18,14 +18,11 @@ const domain = require('domain');
 
 const modulesList = {
 	blocks: '../submodules/blocks',
-	delegates: '../submodules/delegates',
 	rounds: '../submodules/rounds',
-	loader: '../submodules/loader',
 	multisignatures: '../submodules/multisignatures',
 	peers: '../submodules/peers',
 	signatures: '../submodules/signatures',
 	transactions: '../submodules/transactions',
-	transport: '../submodules/transport',
 	processTransactions: '../submodules/process_transactions',
 };
